Move common functions to dedicated modules
