+++ modules/join-path.js
@@ -1,0 +1,11 @@
+export default function (left, right) {
+  left = (left || '').toString()
+  right = (right || '').toString()
+  if (left.endsWith('/') && right.startsWith('/')) {
+    return left + right.slice(1)
+  }
+  if (!left.endsWith('/') && !right.startsWith('/')) {
+    return left + '/' + right
+  }
+  return left + right
+}
